RAG_APP/Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "sundharesan11",email = "sundharesansk11@gmail.com"}
]
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "openai (>=1.65.5,<2.0.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
